python/requirements.txt
groq
keyboard
sounddevice
numpy
pyautogui
pyperclip
requests
python-dotenv
